refactor: split lifecycle gate runners
